
                  % DOCUMENTATION : CADENA_DE_MARKOV()
            % -------------------------------------------------

\name{ENG_CADENA_DE_MARKOV}
\alias{CADENA_DE_MARKOV_english}
\title{CADENA_DE_MARKOV in ENGLISH}

\usage{
CADENA_DE_MARKOV()   # It has not got any arguments
}

\value{Specifically none, since it only activates the application through an empty function.}

\description{
This tiny App have been created as a tool for all these students who
are curious about the Markov Chain Behaviour, more specifically, its behaviour
within any Discrete Domain. Thanks to this App, students will be able to create
all Markov Chains they want, and then, it also allowed to perform three types of
options, which are ; { M.Chain Summary, M.Chain Probability operations, M.Chain Delete }.
}

\examples{
# CADENA_DE_MARKOV()
}
