
                  % DOCUMENTACIÓN : CADENA_DE_MARKOV()
            % --------------------------------------------------

\name{ESP_CADENA_DE_MARKOV}
\alias{CADENA_DE_MARKOV_español}
\title{CADENA_DE_MARKOV en Español}

\usage{
CADENA_DE_MARKOV()   # No tiene argumentos
}

\value{Específicamente ninguno, ya que sólo activa la aplicación mediante una función vacía.}

\description{
Esta pequeña aplicación ha sido creada como una herramienta para todos aquellos estudiantes que sientan curiosidad por el comportamiento de las Cadenas de Markov, más específicamente, su comportamiento dentro de cualquier dominio discreto. Gracias a esta aplicación, los estudiantes podrán crear todas las Cadenas de Markov que deseen, y después, también se les permitirá realizar tres tipos de opciones, las cuales son; { Resumen de C.Markov, Operaciones de probabilidad de la C.Markov, Eliminación de la C.Markov }.
}

\examples{
# CADENA_DE_MARKOV()
}
