
                      % DOCUMENTAZIONE : CADENA_DE_MARKOV()
                % ---------------------------------------------

\name{IT_CADENA_DE_MARKOV}
\alias{CADENA_DE_MARKOV_italiano}
\title{CADENA_DE_MARKOV in ITALIANO}

\usage{
CADENA_DE_MARKOV()   # Non c'è nessun argomento
}

\value{Nello specifico nessuno, poiché l'applicazione è attivata soltanto attraverso una funzione vuota.}

\description{
Questa piccola App è stata creata come uno strumento per tutti quegli studenti che sono curiosi del comportamento della catena di Markov, più specificatamente, del suo comportamento all'interno di qualsiasi dominio discreto. Grazie a questa App, gli studenti saranno in grado di creare tutte le catene di Markov che desiderano, e poi, si permetterà anche di eseguire tre tipi di opzioni che sono; { M.Chain Riepilogo, Operazioni di probabilità di C.Markov, Eliminazione di C.Markov }.
}

\examples{
# CADENA_DE_MARKOV()
}
